+++ src/Core/View.php
@@ -20,7 +20,7 @@
 		protected $model;
 
 		/** @var string[] */
-		protected $matches = [];
+		protected $slugs = [];
 
 
 		/**
@@ -50,10 +50,10 @@
 		}
 
 		/**
-		 * @param string[] $matches
+		 * @param string[] $slugs
 		 */
-		public function setMatches(array $matches): void {
-			$this->matches = $matches;
+		public function setSlugs(array $slugs): void {
+			$this->slugs = $slugs;
 		}
 
 
@@ -81,8 +81,8 @@
 		/**
 		 * @return string[]
 		 */
-		public function getMatches(): array {
-			return $this->matches;
+		public function getSlugs(): array {
+			return $this->slugs;
 		}
 
 	}